% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_quadratic_difference.R
\name{lin_quad_diff}
\alias{lin_quad_diff}
\title{Difference between linear and quadratic fits at a given x value}
\usage{
lin_quad_diff(x, y, x_of_int)
}
\arguments{
\item{x}{A vector with the x-values of the data set to be fitted. Entries must
be in the same order as the y vector.}

\item{y}{A vector with the y-values of the data set to be fitted. Entries must
be in the same order as the x vector.}

\item{x_of_int}{"x of interest". The x value at which the difference between
the fits is to be calculated.}
}
\value{
The difference between the quadratic and linear fit curves at the given
x_of_int value. The difference is calculated as "quadratic" minus "linear".
}
\description{
This function calculates linear and quadratic fits for a given two-
variable data set, and then calculates the difference between the fit curves at
a specified x value. The difference is calculated as "quadratic" - "linear".
}
\examples{
tib1 <- tibble::tribble(
 ~ one, ~ two,
 1, .8,
 2, 4.1,
 3, 8.8,
 4, 16.3,
 5, 24.7
 )
lin_quad_diff(tib1$one, tib1$two, 2)

tib2 <- tibble::tribble(
 ~ a, ~ b, ~ c,
 -7, 7, 1,
 -5, 6.4, 2,
 -3, 4.1, 3,
 -1, 4, 4,
 1, 2.3, 5,
 3, 1, 6,
 5, -0.7, 7
 )
lin_quad_diff(tib2$a, tib2$b, -2)

}
